requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.5
djangorestframework==3.14.0
Pillow==10.2.0
reportlab==4.0.8
whitenoise==6.6.0
django-admin-interface==0.28.1
django-admin-list-filter-dropdown==1.0.3
# translate number to word
inflect==7.0.0
qrcode ==7.4.2
numpy==1.26.4
qrcode-artistic